+++ auspice_client_customisation/splash.js
@@ -67,13 +67,18 @@ class SplashContent extends React.Component {
 
         <CenterContent>
           <P>
-            Currently supported data formats:
+            Currently supported files:
             <ul>
-              <li>Auspice JSON - see the
-                <a href="https://github.com/nextstrain/augur/blob/master/augur/data/schema-export-v2.json"> JSON schema </a>
-                and the
+              <li>Auspice datasets (a main JSON plus any sidecars). See the
                 <a href="https://nextstrain.org/docs/bioinformatics/introduction-to-augur"> Nextstrain docs </a>
-                for how to run the bioinformatics tools to generate these datasets
+                for how to run the bioinformatics tools to generate these datasets.
+                Note that it's possible to drag on multiple datasets, however at most two will be loaded, and it's not possible to control the ordering of these datasets!
+              </li>
+              <li>A nextstrain narrative and associated datasets (JSONs) - see the
+                <a href="https://docs.nextstrain.org/en/latest/tutorials/narratives-how-to-write.html"> Nextstrain docs </a>
+                for how author a narrative. Each dataset the narrative references should have a filename which is the
+                <a href="https://en.wikipedia.org/wiki/URL"> URL path</a> but with forward slashes replaced with underscores ("/"→"_") and a ".json" suffix.
+                Only one narrative can be dropped on at a time!
               </li>
               <li>
                 A phylogenetic tree in
@@ -103,9 +108,7 @@ class SplashContent extends React.Component {
           <Line/>
 
           <P>
-            {`auspice.us ${version} is built by `}<a href="https://twitter.com/hamesjadfield">james hadfield</a>
-            <br/>
-            {`and uses Auspice ${dependencies.auspice}.`}
+            {`auspice.us ${version} (Auspice ${dependencies.auspice})`}
           </P>
           <NextstrainTitle/>
           <GitHub/>